requirements.txt
Authlib==0.11
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.20
uritemplate==3.0.1
urllib3==1.26.6
URLObject==2.4.3
Werkzeug==2.0.1
